backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
langchain==0.2.6
langchain-community==0.2.6
langchain-ollama==0.1.1
langchain-chroma==0.1.3
chromadb==0.5.3
pydantic==2.7.4
pydantic-settings==2.3.4
python-multipart==0.0.9
httpx==0.27.0
python-dotenv==1.0.1
loguru==0.7.2
tenacity==8.3.0
pypdf==4.2.0
tiktoken==0.7.0